src/requirements.txt
Django==3.2
django-3-jet==1.0.8
djangorestframework==3.12.4
django-filter==2.4.0
django-cors-headers==3.7.0
django-environ==0.4.5
psycopg2-binary==2.8.4
django-dynamic-fixture==3.1.1
gunicorn==20.1.0
drf-yasg==1.20.0
django-crum==0.7.5